Numeric graphs are not drawn when data is not parametric as well as numeric (FIND-1072)
When fetching value details, read "values" when "totalValues" is not present.
More correctly we ought to read values only for numeric/date and total values for parametric, but we don't currently pass in the type information and for the time being we only use this value to check whether there are any results (i.e. whether or not to draw numeric/date graphs in Find)

[rev. matthew.gordon]